added bouncy balls and moving platforms to new terrain
merged into new terrain
added renderbouncies function
ckeared bouncies on quit
added print line when player touches win

+++ src/Block.h
@@ -23,6 +23,6 @@ public:
 
     float time;
 
-    Block(SDL_Rect r, int b, bool m, float s, int t);
+    Block(SDL_Rect r, int b, bool m, float s, float t);
     void updatePosition();
 };
